Move mongo config classes to its own folder
